+++ common/src/main/java/com/selfdot/cobblemontrainers/trainer/Generation5AI.java
@@ -161,24 +161,20 @@ public class Generation5AI implements BattleAI {
             List<BattlePokemon> canSwitchTo = activeBattlePokemon.getActor().getPokemonList().stream()
                 .filter(BattlePokemon::canBeSentOut)
                 .toList();
-            if (canSwitchTo.isEmpty()) {
-                return new DefaultActionResponse();
-            }
+            if (canSwitchTo.isEmpty()) return new DefaultActionResponse();
             if (opponentActiveBattlePokemon.isEmpty() || opponentActiveBattlePokemon.get().getBattlePokemon() == null) {
                 return new SwitchActionResponse(canSwitchTo.get(RANDOM.nextInt(canSwitchTo.size())).getUuid());
             }
             BattlePokemon opponent = opponentActiveBattlePokemon.get().getBattlePokemon();
             BattlePokemon nextPokemon = canSwitchTo.stream()
-                .max(Comparator.comparingDouble(pokemon ->
-                    pokemon.getMoveSet().getMoves().stream().map(move ->
-                        powerAndTypeDamage(
-                            move.getPower(),
-                            move.getType(),
-                            opponent.getOriginalPokemon().getPrimaryType(),
-                            opponent.getOriginalPokemon().getSecondaryType()
-                        )
-                    ).max(Double::compare).orElse(0d)
-                )).get();
+                .max(Comparator.comparingDouble(pokemon -> pokemon.getMoveSet().getMoves().stream().map(move ->
+                    powerAndTypeDamage(
+                        move.getPower(),
+                        move.getType(),
+                        opponent.getOriginalPokemon().getPrimaryType(),
+                        opponent.getOriginalPokemon().getSecondaryType()
+                    )
+                ).max(Double::compare).orElse(0d))).get();
             nextPokemon.setWillBeSwitchedIn(true);
             return new SwitchActionResponse(nextPokemon.getUuid());
         }